Design reply interface.

+++ gamble/src/main/java/com/hannuus/gamble/dao/TopicMapper.java
@@ -10,7 +10,7 @@ public interface TopicMapper {
 
     int deleteByExample(TopicExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Topic record);
 
@@ -20,7 +20,7 @@ public interface TopicMapper {
 
     List<Topic> selectByExample(TopicExample example);
 
-    Topic selectByPrimaryKey(Integer id);
+    Topic selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Topic record, @Param("example") TopicExample example);
